Remove old SSL parser
