--- src/boomerang/ssl/parser/SSLScanner.sed
@@ -1,5 +1,0 @@
-s/int __alloca_free_ptr/static int __alloca_free_ptr/
-s/\[\*yy_cp\]/[(unsigned char)\*yy_cp]/
-s/\[yy_c\]/[(unsigned char)yy_c]/
-s,#include "db/sslscanner.h",#include "sslscanner.h",
-s,#include "c/ansi-c-scanner.h",#include "ansi-c-scanner.h",
