+++ tests/unit-tests/boomerang/ssl/parser/ParserTest.cpp
@@ -10,7 +10,7 @@
 #include "ParserTest.h"
 
 
-#include "boomerang/ssl/parser/SSLParser.h"
+#include "boomerang/ssl/RTLInstDict.h"
 #include "boomerang/ssl/statements/Statement.h"
 #include "boomerang/util/log/Log.h"
 
@@ -21,27 +21,10 @@ void ParserTest::testRead()
 {
     RTLInstDict d(false);
 
+    QVERIFY(d.readSSLFile(BOOMERANG_TEST_BASE "share/boomerang/ssl/x86.ssl"));
     QVERIFY(d.readSSLFile(BOOMERANG_TEST_BASE "share/boomerang/ssl/sparc.ssl"));
-}
-
-
-void ParserTest::testExp()
-{
-    QString   s("*i32* r0 := 5 + 6");
-    Statement *a = SSLParser::parseExp(qPrintable(s));
-
-    QVERIFY(a);
-    QString     res;
-    OStream ost(&res);
-    a->print(ost);
-    QCOMPARE(res, "   0 " + s);
-    QString s2 = "*i32* r[0] := 5 + 6";
-    a = SSLParser::parseExp(qPrintable(s2));
-    QVERIFY(a);
-    res.clear();
-    a->print(ost);
-    // Still should print to string s, not s2
-    QCOMPARE(res, "   0 " + s);
+    QVERIFY(d.readSSLFile(BOOMERANG_TEST_BASE "share/boomerang/ssl/ppc.ssl"));
+    QVERIFY(d.readSSLFile(BOOMERANG_TEST_BASE "share/boomerang/ssl/st20.ssl"));
 }
